+++ src/components/Header/TopMenu.jsx
@@ -3,17 +3,11 @@ import React, { useState } from 'react';
 //css
 import styles from '../../styles/css/pages/TopMenu.module.css';
 import { Link } from 'react-router-dom';
+import TopMenuArr from '../Card/TopMenuArr';
 
 //헤더 상단
 const TopMenu = () => {
   const [currentTab , setTab] = useState(0);
-  
-  const TopMenuArr = [
-    {id: '0', name: '회원가입' , url: '/user/join'},
-    {id: '1', name: '로그인' , url: '/user/login'},
-    {id: '2', name: '주문내역' , url: '/member/order'},
-    {id: '3', name: '고객센터' }
-  ];
 
    const selectTopHandler = (index) => {
      setTab(index);
